requirements.txt
Django==2.2
pytz==2019.1
sqlparse==0.3.0
certifi==2019.3.9
chardet==3.0.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-qr-code==1.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
Markdown==2.6.11
Pillow==5.2.0
psycopg2==2.8.1
PyJWT==1.6.4
python-barcode==0.9.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
